Create a new tab bar in the current browser tab
Create a new tab bar using the popup. The tab bar is inserted in the
browser tab from where is opened the popup. The tab bar is insert in
the browser tab only if it still doesn't have a tab bar.

When the tab bar is created from a new browser tab (a tab from where
the user doesn't have browsed), first is navigated to google.com and
after that, the tab bar is inserted.  It is made of this way because
a new tab is not able to insert scripts.