Change the structure of the project
This change group the components per UI. This does easier focus when
working in a specific IU. All components that belong to an IU are
grouped in the same dir. For example, all components that are used in
the tab bar are grouped in the tab-bar dir.

The same changes were done in the style dir.