Create the namespace storage
This namespace let us import the different classes that are inner of it
of the following way:

  import { TabGroup } from './storage';
  new TabGroup();

Or of the follow:

  import * as storage from './storage';
  new storage.TabGroup();

This let us import the different class of the namespace storage in a
context where we use other class with the same name. As is the case of
the component MainPane, in this component we use the component TabGroup
and too the class TabGroup.