Improve the design of the tab bar and add the Add option
Now the design of the tab bar is a float tab bar of white color,
instead of a tab bar gray anchored at the bottom of the page. Also,
now, when a tab is selected in the tab bar, this tab shows a border
of the color blue in its bottom, instead of changing its background
color to the red color.