Update the selected tab from the background and content script
In the commit 0bf82ae the code used to update the selected tab was
moved from the background script to the content-script, this was
done to can to observe when the title of the page change without
update the page. But this brought problems, sometimes the favicon
of the pages isn't found and this broke the extension. For that
reason now the code to update the selected tab is use in both place.
From the background script is update the favicon, title, and URL of
the selected tab. And from the content-script is observe the changes
in the title of the page to update the title of the selected tab.