Fix a problem that updates all opened tabs bars with the same data
When the browser tab updates, the background script sends a message
to the tab bar of the current page to update the title, favicon, and
URL of the selected tab of it. But this message also was sent to the
tabs bars of the other pages, updating the other tabs bars with the
same data of the tab bar of the current page.

This problem was going on for a bug in the chrome.tabs.sendMessage
method. This method must send the message to a specific tab, but it
sends the message to all tabs. This problem is explained better in
the question of Stack Overflow:

https://stackoverflow.com/questions/32508008/chrome-extension-chrome-tabs-sendmessage-is-sending-the-message-to-all-tabs-eve

This problem was solved comparing the id received in the message with
the id of the tab group in the tabs bars. The data of the tab bar is
updated just if both ids are equals.

Other changes:

Improve the code of the tab bar.