Isolate the tab bar using CSS instead of iframe
We decided to change the iframe-isolation for the CSS-isolation for
the reason that the iframe-isolation loads slower the tab bar and
complex the interaction between the page and the tab bar in the iframe.
The CSS-insolation works setting the values of all CSS properties that
we use to its initials values in the first selector of the tab bar.