Update the code that updates the data of the selected tab
From the background script, the message that is sent to the
content-script to update the data of the selected tab now is
sent while the browser tab is loading, before, this message
only was send when the browser tab was completed loaded. We
sent the message while the browser tab is loading to make
sure that the data of the selected tab be updated. And also,
we only are updating the URL and favicon of the selected tab.