Change the way to select the tabs
Previously, the tabs were selected when the user clicks them, and it
keeping of that way yet when the user navigates to another page. When
the user navigated to another page, the title, favicon, and URL of the
selected tab were updated.

Now, when the user navigates to another page, if the selected tab has
the same URL of the page, the selected tab is keep selected. If the
selected tab has a different URL, the tab is deselected. And now, the
title, URL, and favicon of the selected tab there aren't updated when
the user navigates to another page.