Replace the style-loader with the mini-css-extract-plugin
mini-css-extract-plugin let us access the resources of the
extension using the URI chrome-extension://__MSG_@@extension_id__

Update the html-webpack-plugin to the version 5.0.0-beta.1. This
change was made to solve the following bug. The HTML files
generated by the html-webpack-plugin were being deleted each time
that the project was build in watch mode.